+++ lua/scripts/src/wesenGemaMod/Outputs/Helpers/NestedTableConverter.lua
